requirements.txt
allauth-jinja==0.0.3
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==43.0.0
defusedxml==0.8.0rc2
distlib==0.3.8
Django==5.0.6
django-allauth==0.56.1
django-cleanup==8.1.0
django-cors-headers==4.4.0
django-crispy-forms==2.3
django-extensions==3.2.3
django-filter==24.2
django-jinja==2.11.0
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-nested-routers==0.94.1
drf-yasg==1.21.7
et-xmlfile==1.1.0
filelock==3.14.0
graphviz==0.20.3
idna==3.8
inflection==0.5.1
iniconfig==2.0.0
Jinja2==3.1.4
kombu==5.4.2
MarkupSafe==2.1.5
oauthlib==3.2.2
openpyxl==3.1.5
packaging==24.1
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
prompt_toolkit==3.0.48
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.22
pydot==2.0.0
PyJWT==2.9.0
pyparsing==3.1.2
pytest==8.3.3
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
redis==5.0.8
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
virtualenv==20.26.2
waitress==3.0.0
wcwidth==0.2.13
whitenoise==6.7.0
